Add `Weights.merge`
This adds a `merge` method to the weights module, which allows combining
multiple weights together. If any of the weights contradict each other
(by specifying different weights for a given address), an error will be
thrown. The method is thoroughly unit tested.

This is part of work on #1557 (we'll want to be able to merge
`WeightedGraph`s.)

Test plan: Run `yarn test`