Rename `Weights.defaultWeights` to `Weights.empty` (#1589)
`empty` is a more descriptive name for a `Weights` object that has no
weights set, rather than `defaultWeights`.

In every case where we were importing `defaultWeights` as a direct
symbol, I switched to importing the whole module, as usage of
`Weights.empty` makes it clear that the empty object returned will be an
empty weights (as opposed to an empty list or some other empty type).
This is as proposed in the reviews from #1538.

Test plan: It's just a rename and change in imports, so `yarn flow` would
catch any error. `yarn test` passes.

Note: this is a port of #1583, which merged to the wrong branch.