Initiatives: add initiativesMap type
Based on forum discussion, Initiatives should be tracked as in files.

There's one main issue with storing the existing Initiative type as
JSON in a file, there's no natural NodeAddress for a file-based tracker.
This type resolves that by using the keys of a JSON object as a unique
reference within that file.