Initiatives: add conversion functions for InitiativesMap
These functions provide a general purpose conversion between
InitiativesMap and $ReadOnlyArray<Initiative>.

By normalizing the format of the `entryKey` we produce, it allows for a
lot of operations, without needing to come up with your own unique IDs.
Such as merging several InitiativesMaps or loading Initiatives from a
different source, such as Discourse, to store them in files instead.