Create `core/credGraph`
`core/credGraph` is a new module which allows you to compute a
`CredGraph`, which, as the name suggests, contains a `WeightedGraph`,
`TimelineCredScores`, and the `CredParameters` used to generate it.

Thus, you can think of this as a burgeoning replacement for the (ugly
and non-core) TimelineCred module inside `analysis/`.

I intend for this to be the canonical module for generating and
representing cred scores. In followon PRs, I will migrate the
`timelineCred` module to depend on `credGraph`, and will "dissolve" all
of its non-trivial functionality into `credGraph`; e.g. `credGraph` will
have methods for retrieving nodes and edges along with relevant cred
scores and weights.

In the future, I expect we'll also add well-tested and well documented
methods for analyzing cred scores, e.g. for computing cred
decompositions in the style of the `pagerankNodeDecomposition` module,
except with much cleaner APIs that are consistent with the rest of core.

You'll know that this refactor is going well when I entirely remove the
`src/analysis` module. :)

Test plan: Review the code, and observe that I've included smoke unit
tests. `yarn test` passes.