Remove plugins from TimelineCred
This commit removes any dependency on plugins from the TimelineCred
module. Instead, TimelineCred just needs the array of "scoring node
prefixes", i.e. the prefix addresses for all nodes that we will use to
normalize cred totals.

This is not the end of my cred-pipeline refactors (and I actually intend
to eventually remove the TimelineCred module entirely), but it is a nice
waypoint, since we can now close #1557.

Test plan: `yarn test --full` passes.