add initial Grain events and ledger
This commit adds a GrainEvent type, which represents any valid event in
the Grain ledger. Currently, distributions (minting grain) and transfers
are supported. We expect to add more events later (like
boosts/sponsorship).

Given a sequence of GrainEvents, we can construct a Ledger, which keeps
track of the balances and earnings for each actor, and allows retreiving
the events. Currently, we just have an InMemoryLedger, but we may expect
more implementations in the future.

For now, I've kept the interface simple and synchronous, since in the
near-term we will be loading all events in memory. In the future, we may
refactor to an async/generator approach.

We track all grainholders by their node address.

Test plan: Inspect unit tests; `yarn test` passes.