Discourse: add index to database for `findPostInTopic` performance
By stress testing the Discourse plugin, it turns out `findPostInTopic`
is a major bottleneck when creating the graph. It's used in two cases.
When a post has a `replyToPostIndex` and for URLs that reference posts.

The test case I used was https://users.rust-lang.org.
Which had about 110k posts, this query was done about 94.5k times.
Without this index that took in the 10s of minutes. Now less than 10s.

# Conflicts:
#	src/plugins/discourse/mirrorSchema.js