Add a mock middleware module for the grain ui
This contains a skeleton class (interface really) for the data provider
for the upcoming grain user interface. By agreeing on the interface, we
allow work on the backend and frontend to proceed in parallel.

Paired with @ianjdarrow

Test plan: Only mock tests were added, so really `yarn flow` and `yarn
lint` are doing all the work. They pass. :)