Add `legacyConverter` module for the grain ledger
As part of the CredSperiment, we currently track grain distributions and
transfers in a very ad-hoc format in [sourcecred/cred][1]. This commit
adds a formal specification of that ad-hoc type, and tested logic for
converting from that type to our new event format. Once we complete this
conversion, we can throw this code away.

Test plan: `yarn test`, since I added nice unit tests.