Initiatives: add NodeEntry types (#1753)
This is the generalized type that allows us to define contributions to an
Initiative from the same JSON file as the Initiative. See #1739.

The types distinguish between what a user is expected to enter and what
this is internally normalized to. The normalization logic is implemented
in a follow-up PR.