Initiatives: add declaration for NodeEntry types and edges (#1757)
Adds a new NodeType for each NodeEntryField. Allowing multipliers to be
set per field.

Since the "contributes to" edge is very generic and created a naming
conflict, this includes a slightly awkward "contributes to entry" addition.
Including "entry" to differentiate from the existing edge type.

The edge weights for this loosely follow the current edge weight rationale.
When we've merged enough functionality to do in-browser testing of different
weights with an example scenario, I think we should revisit them.